fix: Update repository links (#1190)
Change all mentions of the old repo
(https://github.com/iendeavor/import-meta-env) to the new one.

The best way to do this is to do a proper transfer of the old report to
the new one using Github so that redirects get setup automatically. But
I don't think that's possible retrospectively?